projet/pyproject.toml
[tool.poetry]
name = "visnu_wheel"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.8"
opencv-python = "^4.5.1"
numpy = "^1.20.1"
matplotlib = "^3.3.4"
Kivy = "^2.0.0"
black = "^20.8b1"
mypy = "^0.812"
python-uinput = "^0.11.2"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
